+++ src/components/ui/index.ts
@@ -1,1 +1,2 @@
 export { ParallaxBackground } from "./ParallaxBackground";
+export { GlobalParallaxBackground } from "./GlobalParallaxBackground";